:truck: `App.tsx`에서 `queryClient` 분리
