--- src/features/api/constants.ts
@@ -1,1 +1,0 @@
-export const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
